Bug Fixes! Resolved an issue where the Avatar was falling through the floor while in Charlie_Falling state, because the avatar was ADJACENT to the floor tile and not intersecting it.

+++ Tunnelour/Tunnelour/include/Avatar_Controller.h
@@ -143,7 +143,7 @@ class Avatar_Controller: public Controller,
   std::vector<Tile_Bitmap*> m_wall_tiles;
   std::vector<Tile_Bitmap*> m_ledge_tiles;
 
-  int m_y_fallen;
+  float m_y_fallen;
 
   int m_distance_traveled;
